{{cookiecutter.app_name}}/requirements/dev.txt
# Everything the developer needs in addition to the production requirements
-r prod.txt

# Testing
nose
Flask-Testing
factory-boy>=2.2.1
Flask-WebTest>=0.0.6

# Management script
Flask-Script

# Debug toolbar
Flask-DebugToolbar
